requirements.txt
# Core LangGraph / LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Google Gemini
google-generativeai>=0.8.0
google-genai>=1.0.0

# Supabase
supabase>=2.0.0

# Database connectors
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
duckdb>=1.0.0

# Data & stats
pandas>=2.0.0
numpy>=1.26.0

# Export & utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Web server
fastapi>=0.100.0
uvicorn>=0.20.0

# Neo4j
neo4j>=5.0.0

# GitHub
httpx>=0.27.0

# Async support
anyio>=4.0.0
